% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit_oid.R
\name{cit_oid}
\alias{cit_oid}
\title{Get Citoid data}
\usage{
cit_oid(id, format = "mediawiki", accept = "json",
  accept_language = NULL, ...)
}
\arguments{
\item{id}{(character) id of an article, DOI, ISBN, PMCID or PMID}

\item{format}{(character) the format to use for the resulting citation data.
one of mediawiki (default), zotero, mediawiki-basefields, bibtex}

\item{accept}{(character) json, bibtex, or problem-json}

\item{accept_language}{(character) a field_name; all the rows that have an empty
value in the field_name specified are removed from the result set}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
data.frame, see http://opencitations.net/index/coci/api/v1 for
explanation of the resulting columns
}
\description{
Get Citoid data
}
\examples{
\dontrun{
doi1 <- "10.1108/jd-12-2013-0166"
pmid1 <- 30446726
pmcid1 <- "PMC4679344"
isbn1 <- 1439895619

# different formats
cit_oid(doi1)
cit_oid(pmid1, format = "mediawiki")
cit_oid(pmid1, format = "zotero")
cit_oid(pmid1, format = "mediawiki-basefields")
cat(cit_oid(pmid1, format = "bibtex"))

# PMID example
cit_oid(pmid1, verbose = TRUE)

# ISBN example
cit_oid(isbn1, verbose = TRUE)

# PMCID example
cit_oid(pmcid1)
}
}
\references{
https://en.wikipedia.org/api/rest_v1/#!/Citation/getCitation,
https://www.mediawiki.org/wiki/Citoid
}
