% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcitoid-package.R
\docType{package}
\name{rcitoid-package}
\alias{rcitoid-package}
\alias{rcitoid}
\title{Client for the Open Citations Corpus \url{http://opencitations.net/}}
\description{
Client for the Open Citations Corpus \url{http://opencitations.net/}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
