% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcitoid-package.R
\docType{package}
\name{rcitoid-package}
\alias{rcitoid-package}
\alias{rcitoid}
\title{Client for Citoid (\url{https://www.mediawiki.org/wiki/Citoid})}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
